+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapterTest.kt
@@ -74,6 +74,7 @@ class BitmovinSdkAdapterTest {
             BitmovinSdkAdapter(
                 player,
                 mockk(relaxed = true),
+                mockk(relaxed = true),
                 playerStateMachine,
                 mockk(relaxed = true),
                 mockk(relaxed = true),