+++ collector-exoplayer/src/androidTest/java/com/bitmovin/analytics/exoplayer/PhoneBasicScenariosTest.kt
@@ -143,7 +143,7 @@ class PhoneBasicScenariosTest {
             DataVerifier.verifyStartupSample(eventDataList[0])
             DataVerifier.verifyInvariants(eventDataList)
             DataVerifier.verifyVideoStartEndTimesOnContinuousPlayback(eventDataList)
-            DataVerifier.verifyPlayerSetting(eventDataList, PlayerSettings(true))
+            DataVerifier.verifyPlayerSetting(eventDataList, PlayerSettings(true, true))
         }
     }
 
@@ -271,7 +271,7 @@ class PhoneBasicScenariosTest {
             DataVerifier.verifyStartupSample(eventDataList[0])
             DataVerifier.verifyInvariants(eventDataList)
             DataVerifier.verifyVideoStartEndTimesOnContinuousPlayback(eventDataList)
-            DataVerifier.verifyPlayerSetting(eventDataList, PlayerSettings(true))
+            DataVerifier.verifyPlayerSetting(eventDataList, PlayerSettings(true, true))
         }
 
     @Test
@@ -393,7 +393,7 @@ class PhoneBasicScenariosTest {
             DataVerifier.verifyMpdSourceUrl(eventDataList, liveSample.mpdUrl!!)
             DataVerifier.verifyStartupSample(eventDataList[0])
             DataVerifier.verifyVideoStartEndTimesOnContinuousPlayback(eventDataList)
-            DataVerifier.verifyPlayerSetting(eventDataList, PlayerSettings(false))
+            DataVerifier.verifyPlayerSetting(eventDataList, PlayerSettings(false, true))
             DataVerifier.verifyInvariants(eventDataList)
 
             val filteredList = EventDataUtils.filterNonDeterministicEvents(eventDataList)