+++ collector-theoplayer/src/main/java/com/bitmovin/analytics/theoplayer/TheoPlayerCollector.kt
@@ -60,6 +60,7 @@ internal class TheoPlayerCollector(analyticsConfig: AnalyticsConfig, context: Co
         val playerStatisticsProvider = PlayerStatisticsProvider()
         return TheoPlayerSdkAdapter(
             player,
+            playerContext,
             config,
             stateMachine,
             featureFactory,